clean up neutron_interchain_queries contract dependencies
