+++ artifacts/checksums.txt
@@ -1,3 +1,3 @@
 b8b8650aa62de7172b201b2a8e4e9df37757cb634dbf2e4a7f9b9e353ba5ea46  ibc_transfer.wasm
-a0c59cc8b9946737a36482120308f5fed541bc8012e474d08e24c1009f41751f  neutron_interchain_queries.wasm
+e6b6726ec940b64ae1dd3d04df4ea10f6cd025c27003ffe806d3c2a9f35b9407  neutron_interchain_queries.wasm
 f749ad20727695ec8745dbdf7a23b818cff0b36aeee4be3381192c490be12943  reflect.wasm
