+++ contracts/neutron_interchain_queries/examples/schema.rs
@@ -17,7 +17,7 @@ use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use interchain_queries::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use neutron_interchain_queries::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
