+++ contracts/neutron_interchain_queries/src/lib.rs
@@ -15,6 +15,7 @@
 #![warn(clippy::unwrap_used, clippy::expect_used)]
 
 pub mod contract;
+pub mod msg;
 
 #[allow(clippy::unwrap_used)]
 #[cfg(test)]
