+++ packages/interchain_queries/src/register_queries.rs
@@ -3,7 +3,6 @@ use crate::helpers::{
     create_account_balances_prefix, create_delegation_key, create_params_store_key,
     create_validator_key, decode_and_convert,
 };
-use crate::sudo::TransferRecipientQuery;
 use crate::types::{
     QueryType, BANK_STORE_KEY, KEY_BOND_DENOM, PARAMS_STORE_KEY, STAKING_STORE_KEY,
 };
@@ -12,6 +11,14 @@ use neutron_bindings::msg::NeutronMsg;
 use neutron_bindings::query::InterchainQueries;
 use neutron_bindings::types::{KVKey, KVKeys};
 use schemars::_serde_json::to_string;
+use serde::{Deserialize, Serialize};
+
+/// TransferRecipientQuery represents the request model for transfers query registration.
+#[derive(Serialize, Deserialize)]
+pub struct TransferRecipientQuery {
+    #[serde(rename = "transfer.recipient")]
+    pub recipient: String,
+}
 
 #[allow(clippy::too_many_arguments)]
 /// Registers an interchain query with provided params
